Use the remote authorization service

+++ handlers/delete_resource.go
@@ -1,28 +1,38 @@
 package handlers
 
 import (
+	"log"
+
 	"github.com/go-openapi/runtime/middleware"
+	"github.com/sul-dlss-labs/taco/authorization"
 	"github.com/sul-dlss-labs/taco/db"
 	"github.com/sul-dlss-labs/taco/generated/restapi/operations"
 	"github.com/sul-dlss-labs/taco/storage"
 )
 
 // NewDeleteResource -- Accepts requests to remove a resource and pushes them to Kinesis.
-func NewDeleteResource(repository db.Database, storage storage.Storage) operations.DeleteResourceHandler {
+func NewDeleteResource(repository db.Database, storage storage.Storage, authService authorization.Service) operations.DeleteResourceHandler {
 	return &deleteResourceEntry{
-		repository: repository,
-		storage:    storage,
+		repository:  repository,
+		storage:     storage,
+		authService: authService,
 	}
 }
 
 type deleteResourceEntry struct {
-	repository db.Database
-	storage    storage.Storage
+	repository  db.Database
+	authService authorization.Service
+	storage     storage.Storage
 }
 
 // Handle the delete entry request. If the resource being deleted is a file,
 // Also delete the associated binary from s3
-func (d *deleteResourceEntry) Handle(params operations.DeleteResourceParams) middleware.Responder {
+func (d *deleteResourceEntry) Handle(params operations.DeleteResourceParams, agent *authorization.Agent) middleware.Responder {
+	if !d.authService.CanDeleteResource(agent.Identifier, params.ID) {
+		log.Printf("Agent %s is not permitted to delete resource %s", agent, params.ID)
+		return operations.NewDeleteResourceUnauthorized()
+	}
+
 	if err := d.DeleteAllVersions(params.ID); err != nil {
 		panic(err)
 	}
